Add new Field Change the type of Images from several Images to one Image

+++ app/Http/Controllers/Api/AdvertisementController.php
@@ -16,7 +16,7 @@ class AdvertisementController extends Controller
 
     public function allAdvertisements()
     {
-        $advertisements = Advertisement::translated()->paginate(10);
+        $advertisements = Advertisement::with('replies')->paginate(10);
         return $this->apiResponse($advertisements, 'All Advertisements', 200);
     }
 
@@ -25,12 +25,13 @@ class AdvertisementController extends Controller
         $validator = Validator::make(
             $request->all(),
             [
-                'text_ar'     => 'required|string|min:30|max:255|unique:advertisement_translations,text',
-                'text_en'     => 'required|string|min:30|max:255|unique:advertisement_translations,text',
-                'user_id'     => 'required|string',
-                'section_id'  => 'required|string',
-                'category_id' => 'required|string',
-                'images.*'    => 'required|image|mimes:jpg,jpeg,png,gif',
+                'text_ar'       => 'required|string|min:30|max:255|unique:advertisement_translations,text',
+                'text_en'       => 'required|string|min:30|max:255|unique:advertisement_translations,text',
+                'user_id'       => 'required|string',
+                'section_id'    => 'required|string',
+                'category_id'   => 'string',
+                'image_name'    => 'required|image|mimes:jpg,jpeg,png,gif',
+                'price'         => 'required',
             ]
         );
 
@@ -38,10 +39,14 @@ class AdvertisementController extends Controller
             return $this->apiResponse(null, $validator->errors(), 400);
         }
 
+        $image_name = $this->saveImage($request->file('image_name'), 'images/Advertisements', 600, 315);
+
         $data = [
-            'user_id'             => $request['user_id'],
-            'section_id'          => $request['section_id'],
-            'category_id'         => $request['category_id'],
+            'image_name'  => $image_name,
+            'price'       => $request['price'],
+            'user_id'     => $request['user_id'],
+            'section_id'  => $request['section_id'],
+            'category_id' => $request['category_id'],
             'ar' => [
                 'text'    => $request['text_ar'],
             ],
@@ -50,12 +55,7 @@ class AdvertisementController extends Controller
             ],
         ];
         $advertisement = Advertisement::create($data);
-        $Images = $request->file('images');
-        if ($Images) {
-            foreach ($Images as $Image) {
-                $this->saveAdvertisementImage($Image, 'images/Advertisements', 600, 315, $advertisement);
-            }
-        }
+
         if ($advertisement) {
             return $this->apiResponse($advertisement, 'Advertisement Added Successfully', 200);
         }
@@ -64,25 +64,25 @@ class AdvertisementController extends Controller
 
     public function showAdvertisement($advertisement_id)
     {
-        $advertisement = Advertisement::find($advertisement_id);
+        $advertisement = Advertisement::with('replies')->find($advertisement_id);
         if ($advertisement) {
             return $this->apiResponse($advertisement, 'This Is The Required Advertisement', 200);
         }
         return $this->apiResponse($advertisement, 'The Advertisement Not Found', 401);
     }
 
-    public function updateAdvertisement(Request $request, $advertisement_id)
+    public function updateAdvertisement(Request $request, $id)
     {
-        $advertisement  = Advertisement::find($advertisement_id);
         $validator = Validator::make(
             $request->all(),
             [
-                'text_ar'               => 'required|string|min:30|max:255|unique:advertisement_translations,text,' . $advertisement_id,
-                'text_en'               => 'required|string|min:30|max:255|unique:advertisement_translations,text,' . $advertisement_id,
-                'user_id'               => 'required|string',
-                'section_id'            => 'required|string',
-                'category_id'           => 'required|string',
-                'images.*'              => 'required|image|mimes:jpg,jpeg,png,gif',
+                'text_ar'       => 'required|string|min:30|max:255|unique:advertisement_translations,text,' . $id,
+                'text_en'       => 'required|string|min:30|max:255|unique:advertisement_translations,text,' . $id,
+                'user_id'       => 'required|string',
+                'section_id'    => 'required|string',
+                'category_id'   => 'required|string',
+                'image_name'    => 'required|image|mimes:jpg,jpeg,png,gif',
+                'price'         => 'required',
             ]
         );
 
@@ -90,26 +90,22 @@ class AdvertisementController extends Controller
             return $this->apiResponse(null, $validator->errors(), 400);
         }
 
-        // Delete the Image From Public Folder
-        $Images = $advertisement->images()->pluck('image_name');
-        foreach ($Images as $Image) {
-            $destination = $Image;
+        $advertisement  = Advertisement::find($id);
+        if ($request->hasFile('image_name')) {
+            $destination = $advertisement->image_name;
             if (File::exists($destination)) {
                 File::delete($destination);
             }
-        }
-
-        // Update The Image
-        if ($Images = $request->file('images')) {
-            foreach ($Images as $Image) {
-                $this->saveAdvertisementImage($Image, 'images/Advertisements', 600, 315, $advertisement);
-            }
+            $photo = $request->file('image_name');
+            $image_name = $this->saveImage($photo, 'images/Advertisements', 600, 315);
         }
 
         $data = [
-            'user_id'             => $request['user_id'],
-            'section_id'          => $request['section_id'],
-            'category_id'         => $request['category_id'],
+            'image_name'  => $image_name,
+            'price'       => $request['price'],
+            'user_id'     => $request['user_id'],
+            'section_id'  => $request['section_id'],
+            'category_id' => $request['category_id'],
             'ar' => [
                 'text'    => $request['text_ar'],
             ],
@@ -127,12 +123,9 @@ class AdvertisementController extends Controller
     public function removeAdvertisements($id)
     {
         $advertisement = Advertisement::find($id);
-        $Images = $advertisement->images()->pluck('image_name');
-        foreach ($Images as $Image) {
-            $destination = $Image;
-            if (File::exists($destination)) {
-                File::delete($destination);
-            }
+        $destination = $advertisement->image_name;
+        if (File::exists($destination)) {
+            File::delete($destination);
         }
         $advertisement->delete();
         if ($advertisement) {
@@ -141,9 +134,9 @@ class AdvertisementController extends Controller
         return $this->apiResponse(null, 'Advertisement was not Deleted Successfully', 400);
     }
 
-    public function myAdvertisement($user_id)
+    public function myAdvertisement(Request $request)
     {
-        $my_advertisements = Advertisement::where('user_id', '=', $user_id)->get();
+        $my_advertisements = Advertisement::where('user_id', '=', $request->user_id)->get();
         if (count($my_advertisements) == null) {
             return $this->apiResponse($my_advertisements, 'No ads for you yet', 200);
         }
